follow code design
